requirements.txt
appdirs==1.4.4
asgiref==3.3.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
Django==3.1.3
django-cors-headers==3.5.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
filelock==3.0.12
Flask==1.1.2
google-api-core==1.22.1
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
install==1.3.3
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauth2client==4.1.3
openapi-codec==1.3.2
prettytable==0.7.2
protobuf==3.13.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.15.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.30
Werkzeug==1.0.1
youtube3==1.2.1
